Refactor: move StorageScheduler to dedicated module
